Run Electron with --no-sandbox
As a temporary workaround for no upstream-supported way for Electron to
request nested sandboxes be created by Flatpak, use a version of
https://github.com/flathub/com.unity.UnityHub/blob/master/patch-resources.py
to patch the app resource file and remove the JavaScript code that makes
Skype exit if you pass --no-sandbox. After doing this, we can add
--no-sandbox to our launch script and run newer versions of Skype inside
a Flatpak sandbox. Works around #70 for now.