corrected flaw in codes logic
Problem solved: 
On X11 it would still try to use wayland mode.

Reason for problem:  
Check for $XDG_RUNTIME_DIR/$WAYLAND_DISPLAY only works on wayland because only wayland sets the WAYLAND_DISPLAY variable.
So on wayland testing the previous code by disabling the socket in the flatpak works because the variable pointing to it still exists.

Problem solved by:
By changing the test to XDG_SESSION_TYPE fallback to x11-mode on wayland desktop will now be more complicated than toggling the the wayland-socket in flatseal. It will now require either setting the XDG_SESSION_TYPE variable to something else or unsetting it for the flatpak.